pyproject.toml
[project]
name = "resultados-de-vendas"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "python-decouple>=3.8",
]
